Feat: Use studio.inkdrive.ink for studio URLs
This commit refactors the application to use the `studio.inkdrive.ink`
subdomain for all studio-related pages.

The changes include:
- A new `studio.py` file with a Flask Blueprint for all studio routes.
- The blueprint is registered with a `subdomain='studio'` parameter.
- The studio routes have been moved from `app.py` to `studio.py`.
- The `SERVER_NAME` configuration has been set in `app.py` to enable
  subdomain routing.
- The `url_for` calls in the templates have been updated to use the
  new blueprint.
