+++ templates/dashboard.html
@@ -207,7 +207,7 @@
 
 <div class="studio-grid">
     <!-- Article Studio (Clickable) -->
-    <a href="{{ url_for('article_studio') }}" class="studio-card">
+    <a href="{{ url_for('studio.article_studio') }}" class="studio-card">
         <span class="material-symbols-outlined studio-icon">article</span>
         <h2 class="studio-title">Article Studio</h2>
         <p class="studio-description">Generate long-form, SEO-optimized articles and thought-leadership content.</p>
@@ -215,7 +215,7 @@
     </a>
 
     <!-- Social & Comms Studio -->
-    <a href="{{ url_for('social_studio') }}" class="studio-card">
+    <a href="{{ url_for('studio.social_studio') }}" class="studio-card">
         <span class="material-symbols-outlined studio-icon">group</span>
         <h2 class="studio-title">Social & Comms Studio</h2>
         <p class="studio-description">Craft engaging social media posts, email campaigns, and ad copy.</p>
@@ -223,7 +223,7 @@
     </a>
 
     <!-- Editing & Refinement Studio -->
-    <a href="{{ url_for('editing_studio') }}" class="studio-card">
+    <a href="{{ url_for('studio.editing_studio') }}" class="studio-card">
         <span class="material-symbols-outlined studio-icon">edit_note</span>
         <h2 class="studio-title">Editing & Refinement Studio</h2>
         <p class="studio-description">Improve existing text by changing tone, style, and fixing errors.</p>
@@ -231,7 +231,7 @@
     </a>
 
     <!-- Content Repurposing Studio -->
-    <a href="{{ url_for('repurpose_studio') }}" class="studio-card">
+    <a href="{{ url_for('studio.repurpose_studio') }}" class="studio-card">
         <span class="material-symbols-outlined studio-icon">transform</span>
         <h2 class="studio-title">Content Repurposing Studio</h2>
         <p class="studio-description">Turn one piece of content into many, like an article into a Twitter thread.</p>
@@ -239,7 +239,7 @@
     </a>
 
     <!-- SEO Strategy Studio -->
-    <a href="{{ url_for('seo_studio') }}" class="studio-card">
+    <a href="{{ url_for('studio.seo_studio') }}" class="studio-card">
         <span class="material-symbols-outlined studio-icon">trending_up</span>
         <h2 class="studio-title">SEO Strategy Studio</h2>
         <p class="studio-description">Generate keywords, analyze headlines, and build a content strategy.</p>
@@ -247,7 +247,7 @@
     </a>
 
     <!-- Brainstorming Studio -->
-    <a href="{{ url_for('brainstorming_studio') }}" class="studio-card">
+    <a href="{{ url_for('studio.brainstorming_studio') }}" class="studio-card">
         <span class="material-symbols-outlined studio-icon">lightbulb</span>
         <h2 class="studio-title">Brainstorming Studio</h2>
         <p class="studio-description">Generate new ideas, titles, and outlines for your content.</p>
@@ -255,7 +255,7 @@
     </a>
 
     <!-- Scriptwriting Studio -->
-    <a href="{{ url_for('scriptwriting_studio') }}" class="studio-card">
+    <a href="{{ url_for('studio.scriptwriting_studio') }}" class="studio-card">
         <span class="material-symbols-outlined studio-icon">movie</span>
         <h2 class="studio-title">Scriptwriting Studio</h2>
         <p class="studio-description">Generate scripts for YouTube, TikTok, and other video platforms.</p>
@@ -263,7 +263,7 @@
     </a>
 
     <!-- E-commerce Studio -->
-    <a href="{{ url_for('ecommerce_studio') }}" class="studio-card">
+    <a href="{{ url_for('studio.ecommerce_studio') }}" class="studio-card">
         <span class="material-symbols-outlined studio-icon">storefront</span>
         <h2 class="studio-title">E-commerce Studio</h2>
         <p class="studio-description">Write compelling product descriptions and marketing copy.</p>
@@ -271,7 +271,7 @@
     </a>
 
     <!-- Web Copy Studio -->
-    <a href="{{ url_for('webcopy_studio') }}" class="studio-card">
+    <a href="{{ url_for('studio.webcopy_studio') }}" class="studio-card">
         <span class="material-symbols-outlined studio-icon">web</span>
         <h2 class="studio-title">Web Copy Studio</h2>
         <p class="studio-description">Craft landing page copy, ad copy, and more using proven frameworks.</p>
@@ -279,7 +279,7 @@
     </a>
 
     <!-- Business Docs Studio -->
-    <a href="{{ url_for('business_studio') }}" class="studio-card">
+    <a href="{{ url_for('studio.business_studio') }}" class="studio-card">
         <span class="material-symbols-outlined studio-icon">business_center</span>
         <h2 class="studio-title">Business Docs Studio</h2>
         <p class="studio-description">Generate professional press releases, job descriptions, and reports.</p>
